+++ crates/sncast/src/lib.rs
@@ -45,6 +45,8 @@ use thiserror::Error;
 use url::Url;
 
 pub mod helpers;
+#[path = "starknet_commands/ledger/mod.rs"]
+pub mod ledger;
 pub mod response;
 pub mod state;