+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Cast
+
+#### Added
+
+- `sncast ledger` command for simple Ledger operations.
+
 ## [0.57.0] - 2026-02-24
 
 ### Forge